Implements resolveWorkspaceSymbol as command
