+++ org.eclipse.jdt.ls.core/src/org/eclipse/jdt/ls/core/internal/JDTDelegateCommandHandler.java
@@ -23,12 +23,13 @@ import org.eclipse.jdt.ls.core.internal.commands.DiagnosticsCommand;
 import org.eclipse.jdt.ls.core.internal.commands.OrganizeImportsCommand;
 import org.eclipse.jdt.ls.core.internal.commands.ProjectCommand;
 import org.eclipse.jdt.ls.core.internal.commands.ProjectCommand.ClasspathOptions;
+import org.eclipse.jdt.ls.core.internal.commands.SourceAttachmentCommand;
+import org.eclipse.jdt.ls.core.internal.commands.TypeHierarchyCommand;
 import org.eclipse.jdt.ls.core.internal.handlers.FormatterHandler;
 import org.eclipse.jdt.ls.core.internal.handlers.ResolveSourceMappingHandler;
 import org.eclipse.jdt.ls.core.internal.managers.GradleProjectImporter;
-import org.eclipse.jdt.ls.core.internal.commands.SourceAttachmentCommand;
-import org.eclipse.jdt.ls.core.internal.commands.TypeHierarchyCommand;
 import org.eclipse.lsp4j.ResolveTypeHierarchyItemParams;
+import org.eclipse.lsp4j.SymbolInformation;
 import org.eclipse.lsp4j.TextDocumentPositionParams;
 import org.eclipse.lsp4j.TypeHierarchyDirection;
 import org.eclipse.lsp4j.TypeHierarchyItem;
@@ -118,6 +119,9 @@ public class JDTDelegateCommandHandler implements IDelegateCommandHandler {
 				case "java.project.upgradeGradle":
 					String projectUri = (String) arguments.get(0);
 					return GradleProjectImporter.upgradeGradleVersion(projectUri, monitor);
+				case "java.project.resolveWorkspaceSymbol":
+					SymbolInformation si = JSONUtility.toModel(arguments.get(0), SymbolInformation.class);
+					return ProjectCommand.resolve(si);
 				default:
 					break;
 			}
