+++ org.eclipse.jdt.ls.core/src/org/eclipse/jdt/ls/core/internal/handlers/WorkspaceSymbolHandler.java
@@ -12,22 +12,14 @@
  *******************************************************************************/
 package org.eclipse.jdt.ls.core.internal.handlers;
 
-import java.util.ArrayDeque;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Deque;
 import java.util.List;
-import java.util.Objects;
 
 import org.apache.commons.lang3.StringUtils;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.OperationCanceledException;
 import org.eclipse.jdt.core.Flags;
-import org.eclipse.jdt.core.IJavaElement;
 import org.eclipse.jdt.core.IJavaProject;
-import org.eclipse.jdt.core.IParent;
-import org.eclipse.jdt.core.IType;
-import org.eclipse.jdt.core.ITypeRoot;
 import org.eclipse.jdt.core.JavaModelException;
 import org.eclipse.jdt.core.search.IJavaSearchConstants;
 import org.eclipse.jdt.core.search.IJavaSearchScope;
@@ -41,7 +33,6 @@ import org.eclipse.jdt.ls.core.internal.JDTUtils;
 import org.eclipse.jdt.ls.core.internal.JavaLanguageServerPlugin;
 import org.eclipse.jdt.ls.core.internal.ProjectUtils;
 import org.eclipse.jdt.ls.core.internal.preferences.PreferenceManager;
-import org.eclipse.lsp4j.AbstractWorkDoneProgressOptions;
 import org.eclipse.lsp4j.Location;
 import org.eclipse.lsp4j.SymbolInformation;
 import org.eclipse.lsp4j.SymbolKind;
@@ -101,7 +92,8 @@ public class WorkspaceSymbolHandler {
 							if (Flags.isDeprecated(match.getType().getFlags())) {
 								if (preferenceManager != null && preferenceManager.getClientPreferences().isSymbolTagSupported()) {
 									symbolInformation.setTags(List.of(SymbolTag.Deprecated));
-								} else {
+								}
+								else {
 									symbolInformation.setDeprecated(true);
 								}
 							}
@@ -160,7 +152,8 @@ public class WorkspaceSymbolHandler {
 								if (Flags.isDeprecated(match.getMethod().getFlags())) {
 									if (preferenceManager != null && preferenceManager.getClientPreferences().isSymbolTagSupported()) {
 										symbolInformation.setTags(List.of(SymbolTag.Deprecated));
-									} else {
+									}
+									else {
 										symbolInformation.setDeprecated(true);
 									}
 								}
@@ -217,78 +210,5 @@ public class WorkspaceSymbolHandler {
 			super(query);
 			this.projectName = projectName;
 		}
-	}
-
-	public static class WorkspaceSymbolResolveOptions extends AbstractWorkDoneProgressOptions {
-
-		private Boolean resolveProvider;
-
-		public WorkspaceSymbolResolveOptions(Boolean resolveProvider) {
-			this.resolveProvider = resolveProvider;
-		}
-
-		@Override
-		public int hashCode() {
-			final int prime = 31;
-			int result = super.hashCode();
-			result = prime * result + Objects.hash(resolveProvider);
-			return result;
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if (this == obj) {
-				return true;
-			}
-			if (!super.equals(obj)) {
-				return false;
-			}
-			if (getClass() != obj.getClass()) {
-				return false;
-			}
-			WorkspaceSymbolResolveOptions other = (WorkspaceSymbolResolveOptions) obj;
-			return Objects.equals(resolveProvider, other.resolveProvider);
-		}
-
-	}
-
-	public final static WorkspaceSymbolResolveOptions WORKSPACE_SYMBOL_RESOLVE_OPTIONS = new WorkspaceSymbolResolveOptions(Boolean.TRUE);
-
-	public static SymbolInformation resolve(SymbolInformation request) {
-		ITypeRoot unit = JDTUtils.resolveTypeRoot(request.getLocation().getUri());
-		if (unit == null || !unit.exists()) {
-			return null;
-		}
-		Location location = null;
-		IJavaElement element = null;
-
-		try {
-			Deque<IJavaElement> children = new ArrayDeque<>(Arrays.asList(unit.getChildren()));
-			while (element == null && !children.isEmpty()) {
-				IJavaElement child = children.pop();
-				if (child instanceof IType) {
-
-					if (request.getName().equals(child.getElementName())) {
-						element = child;
-						break;
-					}
-					children.addAll(Arrays.asList(((IParent) child).getChildren()));
-				}
-
-			}
-			location = JDTUtils.toLocation(element);
-
-			SymbolInformation si = new SymbolInformation();
-			si.setName(request.getName());
-			si.setContainerName(request.getContainerName());
-			si.setLocation(location);
-
-			return si;
-
-		} catch (JavaModelException e) {
-			JavaLanguageServerPlugin.logError("Problem resolving symbol information for " + unit.getElementName());
-		}
-
-		return null;
 	}
 }
