--- org.eclipse.jdt.ls.core/src/org/eclipse/jdt/ls/core/internal/handlers/WorkspaceSymbolsService.java
@@ -1,28 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2022 Red Hat Inc. and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License 2.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     Red Hat Inc. - initial API and implementation
- *******************************************************************************/
-package org.eclipse.jdt.ls.core.internal.handlers;
-
-import java.util.concurrent.CompletableFuture;
-
-import org.eclipse.lsp4j.SymbolInformation;
-import org.eclipse.lsp4j.jsonrpc.services.JsonRequest;
-import org.eclipse.lsp4j.jsonrpc.services.JsonSegment;
-
-
-@JsonSegment("workspaceSymbol")
-public interface WorkspaceSymbolsService {
-
-	@JsonRequest
-	public CompletableFuture<SymbolInformation> resolve(SymbolInformation params);
-
-}
